Rewrite a section to tone it down a little.
The original was a little too violatile.

The use of a 'strawman' can be taken as insulting the reader. Even if the
reader does not feel insulted it whips them up to see this argument as
fundamentally wrong, which is not true.

Furthermore, there was random emphasis being put on words, and certain
sections felt a little subjective.

So I rewrote it to be a little less 'art is an explosion' and a little
more 'Bob Ross'.

Truthfully, palette control needs its own section.