KritaFAQ cosmetic fixes.
This is a set of cosmetic fixes (typos, standardization, punctuation, etc.)
found while doing the translation. Note that the changes should not present a content
change on its own, some parts of the sentences that where altered are
simply rephrased, but translations on the works (Chinese/Spanish)
should have no problems as the interpretation still will be the same.

Differential Revision: https://phabricator.kde.org/D13426